Prettier solidity code

+++ contracts/WETH.sol
@@ -6,13 +6,14 @@ import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 
 contract WETH is ERC20 {
     constructor() ERC20("Wrapped Ether", "WETH") {}
-    
+
     function mint() external payable {
-        _mint (msg.sender,msg.value);
+        _mint(msg.sender, msg.value);
     }
+
     // now taking the ether back
-    
-    function burn(uint amount) external {
+
+    function burn(uint256 amount) external {
         _burn(msg.sender, amount);
         (bool success, ) = msg.sender.call{value: amount}("");
         require(success, "Transfer failed.");
